requirements.txt
streamlit
pandas
groq
graphviz
openpyxl
python-pptx
vsdx
xmltodict